requirements.txt
# ============================================
# FRAMEWORK WEB
# ============================================
Flask==3.0.0
Werkzeug==3.0.1

# ============================================
# BANCO DE DADOS - MySQL
# ============================================
mysql-connector-python==8.3.0

# ============================================
# AUTENTICAÇÃO JWT
# ============================================
PyJWT==2.8.0

# ============================================
# SEGURANÇA - Hash de senhas
# ============================================
bcrypt==4.1.2

# ============================================
# VARIÁVEIS DE AMBIENTE
# ============================================
python-dotenv==1.0.0

# ============================================
# CORS (Cross-Origin Resource Sharing)
# ============================================
Flask-CORS==4.0.0

# ============================================
# UTILITÁRIOS
# ============================================
# Para datas e horários (já vem com Python, mas explícito)
python-dateutil==2.8.2

